+++ run_amf.py
@@ -16,7 +16,7 @@ import utils.contextvars_stub  # noqa
 from utils.bootstrap import ensure_deps
 ensure_deps()
 
-import uvicorn
+from utils.serve import serve
 from amf import config
 from amf.main import app
 
@@ -26,5 +26,5 @@ DEFAULT_PORT = config.AMF_PORT
 if __name__ == "__main__":
     host = sys.argv[1] if len(sys.argv) > 1 else DEFAULT_HOST
     port = int(sys.argv[2]) if len(sys.argv) > 2 else DEFAULT_PORT
-    print("Starting AMF service: http://{}:{}".format(host, port))
-    uvicorn.run(app, host=host, port=port)
+    print("Starting AMF service: https://{}:{}".format(host, port))
+    serve("amf.main:app", host=host, port=port)